feat: user-entity added

+++ src/main/java/com/digitaldark/ChambeaPe_Api/repository/UserRepository.java
@@ -5,13 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.digitaldark.ChambeaPe_Api.model.UsersEntity;
 
-public interface UserRepository extends JpaRepository<UsersEntity, Long> {
+public interface UserRepository extends JpaRepository<UsersEntity, Integer> {
 
-    UsersEntity findById(long id);
+    UsersEntity findById(int id);
     UsersEntity findByEmailAndPassword(String email, String password);
     List<UsersEntity> findAll();
 
-    boolean existsById(long id);
+    boolean existsById(int id);
     boolean existsByEmailAndPassword(String email, String password);
     boolean existsByEmailOrPhoneNumber(String email, String phoneNumber);
 }
